Move CLI commands into TypeScript modules
